# heimdall/docs/swagger.yaml
basePath: /
definitions:
...
  forms.ListRunsDto:
    properties:
      next_page_token:
        type: string
      runs:
        items:
          $ref: '#/definitions/forms.RunStatusDto'
        type: array
      total:
        type: integer
    type: object
...
  forms.RunRequestDto:
    properties:
      tags:
        additionalProperties:
          type: string
        type: object
      workflow_engine_parameters:
        additionalProperties:
          type: string
        type: object
      workflow_params:
        additionalProperties: true
        type: object
      workflow_type:
        type: string
      workflow_type_version:
        type: string
      workflow_url:
        type: string
    type: object
  forms.RunStatusDto:
    properties:
      end_time:
        type: string
      id:
        type: string
      request:
        $ref: '#/definitions/forms.RunRequestDto'
      start_time:
        type: string
      state:
        type: string
      user:
        type: string
    type: object
...
  forms.WorkflowDto:
    properties:
      author:
        type: string
      class:
        type: string
      content:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      project_name:
        type: string
      steps:
        items:
          $ref: '#/definitions/forms.WorkflowStepDto'
        type: array
      tags:
        additionalProperties: true
        type: object
      updated_at:
        type: string
    type: object
  forms.WorkflowForm:
    properties:
      class:
        type: string
      content:
        type: string
      description:
        type: string
      name:
        type: string
      project_id:
        type: string
      steps:
        items:
          $ref: '#/definitions/forms.WorkflowStepForm'
        type: array
      tags:
        additionalProperties: true
        type: object
    type: object
...
  forms.WorkflowStepDto:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  forms.WorkflowStepForm:
    properties:
      content:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  forms.WorkflowsDto:
    properties:
      next_page_token:
        type: string
      total:
        type: integer
      workflows:
        items:
          $ref: '#/definitions/forms.WorkflowDto'
        type: array
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Heimdall server.
  license:
    name: VinBigData
    url: http://vinbigdata.org
  termsOfService: http://swagger.io/terms/
  title: Heimdall API
  version: "1.0"
paths:
...
  /workflows:
    get:
      description: Get workflows by param
      parameters:
      - description: page size
        in: query
        name: page_size
        required: true
        type: integer
      - description: page token
        in: query
        name: page_token
        required: true
        type: integer
      - description: Filter. Split by ;
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Workflows information
          schema:
            $ref: '#/definitions/forms.WorkflowsDto'
      summary: Get workflows by param
      tags:
      - workflow
    post:
      description: Create workflow
      parameters:
      - description: workflow info
        in: body
        name: createRequest
        required: true
        schema:
          $ref: '#/definitions/forms.WorkflowForm'
      produces:
      - application/json
      responses:
        "200":
          description: Workflow info
          schema:
            $ref: '#/definitions/forms.WorkflowDto'
      summary: Create workflow
      tags:
      - workflow
  /workflows/:workflow_id:
    delete:
      description: Delete workflow by id
      parameters:
      - description: workflow id
        in: path
        name: workflow_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Delete workflow by id
      tags:
      - workflow
    get:
      description: Get workflows by ID
      parameters:
      - description: workflow id
        in: path
        name: workflow_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: workflow response
          schema:
            $ref: '#/definitions/forms.WorkflowDto'
      summary: Get workflows by ID
      tags:
      - workflow
    put:
      description: Update workflow by id
      parameters:
      - description: workflow id
        in: path
        name: workflow_id
        required: true
        type: integer
      - description: Update info
        in: body
        name: updateRequest
        required: true
        schema:
          $ref: '#/definitions/forms.WorkflowForm'
      produces:
      - application/json
      responses:
        "200":
          description: Workflow updated info
          schema:
            $ref: '#/definitions/forms.WorkflowDto'
      summary: Update workflow by id
      tags:
      - workflow
  /workflows/:workflow_id/runs:
    get:
      description: Get run created from workflow id
      parameters:
      - description: workflow id
        in: path
        name: workflow_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Run info
          schema:
            $ref: '#/definitions/forms.ListRunsDto'
      summary: Get run created from workflow id
      tags:
      - run
swagger: "2.0"